Respond to comments
- Moved MethodFields and its renamed method to method.rs
- Updated the expected outputs
- Refactor the lookup method to be generic, and remove the superfluous trait
