+++ crates/cxx-qt-gen/src/generator/structuring/qobject.rs
@@ -7,7 +7,7 @@ use crate::naming::Name;
 use crate::parser::inherit::ParsedInheritedMethod;
 use crate::parser::method::ParsedMethod;
 use crate::parser::signals::ParsedSignal;
-use crate::parser::{qenum::ParsedQEnum, qobject::ParsedQObject, Invokable};
+use crate::parser::{qenum::ParsedQEnum, qobject::ParsedQObject};
 use proc_macro2::Ident;
 use syn::{Error, Result};
 
@@ -21,10 +21,10 @@ pub struct StructuredQObject<'a> {
     pub signals: Vec<&'a ParsedSignal>,
 }
 
-fn lookup(invokables: &[impl Invokable], id: &Ident) -> Option<Name> {
+fn lookup<T>(invokables: &[T], id: &Ident, name_getter: impl Fn(&T) -> &Name) -> Option<Name> {
     invokables
         .iter()
-        .map(|invokable| invokable.name())
+        .map(name_getter)
         .find(|name| name.rust_unqualified() == id)
         .cloned()
 }
@@ -47,14 +47,14 @@ impl<'a> StructuredQObject<'a> {
 
     /// Returns the name of the method with the provided Rust ident if it exists, or an error
     pub fn method_lookup(&self, id: &Ident) -> Result<Name> {
-        lookup(&self.methods, id)
-            .or_else(|| lookup(&self.inherited_methods, id)) // fallback to searching inherited methods
+        lookup(&self.methods, id, |method| &method.name)
+            .or_else(|| lookup(&self.inherited_methods, id, |inherited| &inherited.name)) // fallback to searching inherited methods
             .ok_or_else(|| Error::new_spanned(id, format!("Method with name '{id}' not found!")))
     }
 
     /// Returns the name of the signal with the provided Rust ident if it exists, or an error
     pub fn signal_lookup(&self, id: &Ident) -> Result<Name> {
-        lookup(&self.signals, id)
+        lookup(&self.signals, id, |signal| &signal.name)
             .ok_or_else(|| Error::new_spanned(id, format!("Signal with name '{id}' not found!")))
     }
