+++ crates/cxx-qt-gen/test_outputs/invokables.rs
@@ -42,46 +42,57 @@ mod ffi {
     }
     extern "Rust" {
         #[cxx_name = "cppMethodWrapper"]
+        #[doc(hidden)]
         fn cpp_method(self: &MyObject);
     }
     extern "Rust" {
         #[cxx_name = "invokableWrapper"]
+        #[doc(hidden)]
         fn invokable(self: &MyObject);
     }
     extern "Rust" {
         #[cxx_name = "invokableMutableWrapper"]
+        #[doc(hidden)]
         fn invokable_mutable(self: Pin<&mut MyObject>);
     }
     extern "Rust" {
         #[cxx_name = "invokableParametersWrapper"]
+        #[doc(hidden)]
         fn invokable_parameters(self: &MyObject, opaque: &QColor, trivial: &QPoint, primitive: i32);
     }
     extern "Rust" {
         #[cxx_name = "invokableReturnOpaqueWrapper"]
+        #[doc(hidden)]
         fn invokable_return_opaque(self: Pin<&mut MyObject>) -> UniquePtr<Opaque>;
     }
     extern "Rust" {
         #[cxx_name = "invokableReturnTrivialWrapper"]
+        #[doc(hidden)]
         fn invokable_return_trivial(self: Pin<&mut MyObject>) -> QPoint;
     }
     extern "Rust" {
         #[cxx_name = "invokableFinalWrapper"]
+        #[doc(hidden)]
         fn invokable_final(self: &MyObject);
     }
     extern "Rust" {
         #[cxx_name = "invokableOverrideWrapper"]
+        #[doc(hidden)]
         fn invokable_override(self: &MyObject);
     }
     extern "Rust" {
         #[cxx_name = "invokableVirtualWrapper"]
+        #[doc(hidden)]
         fn invokable_virtual(self: &MyObject);
     }
     extern "Rust" {
         #[cxx_name = "invokableResultTupleWrapper"]
+        #[doc(hidden)]
         fn invokable_result_tuple(self: &MyObject) -> Result<()>;
     }
     extern "Rust" {
         #[cxx_name = "invokableResultTypeWrapper"]
+        #[doc(hidden)]
         fn invokable_result_type(self: &MyObject) -> Result<String>;
     }
     unsafe extern "C++" {
