+++ crates/cxx-qt-gen/test_outputs/properties.rs
@@ -188,14 +188,17 @@ mod ffi {
     }
     extern "Rust" {
         #[cxx_name = "myGetterWrapper"]
+        #[doc(hidden)]
         fn my_getter(self: &MyObject) -> i32;
     }
     extern "Rust" {
         #[cxx_name = "MyCustomSetterWrapper"]
+        #[doc(hidden)]
         fn my_setter(self: Pin<&mut MyObject>, value: i32);
     }
     extern "Rust" {
         #[cxx_name = "myResetFnWrapper"]
+        #[doc(hidden)]
         fn myResetFn(self: Pin<&mut MyObject>);
     }
     unsafe extern "C++" {
