+++ crates/cxx-qt-gen/test_outputs/signals.rs
@@ -36,6 +36,7 @@ mod ffi {
     }
     extern "Rust" {
         #[cxx_name = "invokableWrapper"]
+        #[doc(hidden)]
         fn invokable(self: Pin<&mut MyObject>);
     }
     unsafe extern "C++" {
